refactor: rename internal/server_test to internal/servertest
Fix package/directory name mismatch — the directory was named server_test
but declared package server. Rename to servertest with matching package
declaration for clarity.
