+++ internal/servertest/health_filter_test.go
@@ -1,4 +1,4 @@
-package server
+package servertest
 
 import (
 	"testing"
