+++ internal/servertest/lb_virtual_validation_test.go
@@ -1,7 +1,7 @@
 //go:build e2e
 // +build e2e
 
-package server
+package servertest
 
 import (
 	"fmt"
