+++ internal/servertest/no_service_available_repro_test.go
@@ -1,4 +1,4 @@
-package server
+package servertest
 
 import (
 	"testing"
